Adds Graph Functions
authorGraph will work for making histograms given the author name, with
number of bins equal to number of months.
filterData is in progress
topicGraph is a placeholder
timeFilter was made to test the time filtering part of filterData. It
can take before, after, or both inputs to filter data from before a
certain day, after a certain day, or in between two days.